Support graphic image content
